+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##
 - Bugfix `LAN_X_LOCO_INFO` DB0 (most significant bits of loco address) ([#11](https://github.com/ZIMO-Elektronik/Z21/issues/11))
+- Bugfix `LAN_RAILCOM_GETDATA` used little endian loco address ([#12](https://github.com/ZIMO-Elektronik/Z21/issues/12))
 
 ## 0.3.1
 - Bugfix `LAN_X_BC_PROGRAMMING_MODE` comes only after `LAN_X_CV_NACK_SC`, `LAN_X_CV_NACK` or `LAN_X_CV_RESULT` ([#7](https://github.com/ZIMO-Elektronik/Z21/issues/7))